backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.2
chromadb==0.5.5
pymupdf==1.24.10
openai==1.47.0
google-generativeai==0.8.2
sentence-transformers==3.1.1
numpy==1.26.4